+++ resolwe/flow/elastic_indexes/__init__.py
@@ -1,0 +1,3 @@
+"""Elastic search indexes for Resolwe models."""
+
+from .data import DataDocument, DataIndex