+++ resolwe/permissions/tests/test_data.py
@@ -48,6 +48,11 @@ class DataTestCase(ResolweAPITestCase):
 
         super(DataTestCase, self).setUp()
 
+        # Reindex data objects as they are loaded in fixtures.
+        # TODO: Remove this when we get rid of fixtures.
+        from resolwe.elastic.builder import index_builder
+        index_builder.build()
+
     def tearDown(self):
         for data in Data.objects.all():
             data_dir = os.path.join(settings.FLOW_EXECUTOR['DATA_DIR'], str(data.id))