+++ docs/CHANGELOG.rst
@@ -15,6 +15,7 @@ Changed
 - **BACKWARD INCOMPATIBLE:** Use Elasticsearch version 5.x
 - **BACKWARD INCOMPATIBLE:** Raise an error if an invalid query argument is
   used in Elasticsearch viewsets
+- **BACKWARD INCOMPATIBLE:** Switch ``Data`` API viewset to use Elasticsearch
 - Terminate the executor if listener response with error message
 - ``verbosity`` setting is no longer propagated to the executor